+++ go-app/cabdebugger/cabdebugger.go
@@ -112,7 +112,13 @@ func (cd *CabDebugger) Start(ctx context.Context) {
 	go func() {
 		socket_channel, unsubscribe_socket_channel := cd.Connector.Subscribe()
 		ticker := time.NewTicker(333 * time.Millisecond)
+		slow_ticker := time.NewTicker(time.Second)
 		for {
+			tick_channel := ticker.C
+			if !cd.TSWAPI.CanConnect() {
+				tick_channel = slow_ticker.C
+			}
+
 			select {
 			case msg := <-socket_channel:
 				if msg.EventName == "current_drivable_actor" && msg.Properties["name"] != cd.State.DrivableActorName {
@@ -120,8 +126,8 @@ func (cd *CabDebugger) Start(ctx context.Context) {
 				}
 				if msg.EventName == "sync_control_value" {
 					control_state, has_control_state := cd.State.Controls.Get(msg.Properties["property"])
-					if cd.TSWAPI.Enabled() && !has_control_state {
-						/* if the API is enabled - it should drive the existance of the controls */
+					if cd.TSWAPI.Enabled() && cd.TSWAPI.CanConnect() && !has_control_state {
+						/* if the API is enabled and connectable - it should drive the existance of the controls */
 						continue
 					}
 
@@ -133,10 +139,11 @@ func (cd *CabDebugger) Start(ctx context.Context) {
 					control_state.CurrentNormalizedValue = current_normalized_value
 					cd.State.Controls.Set(msg.Properties["property"], control_state)
 				}
-			case <-ticker.C:
+			case <-tick_channel:
 				go cd.updateControlStateFromAPI()
 			case <-ctx.Done():
 				ticker.Stop()
+				slow_ticker.Stop()
 				unsubscribe_socket_channel()
 				return
 			}